+++ test/controls/camera-button.spec.js
@@ -31,13 +31,13 @@ describe('controls.CameraButton', function() {
 
         expect(button.el().nodeName).toEqual('BUTTON');
         expect(button.on).toBeFunction();
-        expect(button.enabled_).toEqual(true);
+        expect(button.enabled_).toBeTrue();
         expect(button.controlText_).toEqual('Image');
 
         let styleClasses = ['vjs-camera-button', 'vjs-control', 'vjs-button',
             'vjs-icon-photo-camera'];
         styleClasses.forEach((e) => {
-            expect(button.hasClass(e)).toEqual(true);
+            expect(button.hasClass(e)).toBeTrue();
         });
     });
 
@@ -50,11 +50,11 @@ describe('controls.CameraButton', function() {
     it('changes appearance when startRecord or stopRecord is triggered', function(done) {
         let button = new CameraButton(player);
 
-        expect(button.hasClass('vjs-icon-photo-camera')).toEqual(true);
+        expect(button.hasClass('vjs-icon-photo-camera')).toBeTrue();
 
         player.one('startRecord', function() {
-            expect(button.hasClass('vjs-icon-photo-camera')).toEqual(false);
-            expect(button.hasClass('vjs-icon-replay')).toEqual(true);
+            expect(button.hasClass('vjs-icon-photo-camera')).toBeFalse();
+            expect(button.hasClass('vjs-icon-replay')).toBeTrue();
             expect(button.controlText_).toEqual('Retry');
 
             setTimeout(function() {
@@ -65,8 +65,8 @@ describe('controls.CameraButton', function() {
             }, 2000);
         });
         player.one('stopRecord', function() {
-            expect(button.hasClass('vjs-icon-replay')).toEqual(false);
-            expect(button.hasClass('vjs-icon-photo-camera')).toEqual(true);
+            expect(button.hasClass('vjs-icon-replay')).toBeFalse();
+            expect(button.hasClass('vjs-icon-photo-camera')).toBeTrue();
             expect(button.controlText_).toEqual('Image');
 
             done();