+++ test/controls/record-toggle.spec.js
@@ -25,13 +25,13 @@ describe('controls.RecordToggle', function() {
 
         expect(toggle.el().nodeName).toEqual('BUTTON');
         expect(toggle.on).toBeFunction();
-        expect(toggle.enabled_).toEqual(true);
+        expect(toggle.enabled_).toBeTrue();
         expect(toggle.controlText_).toEqual('Record');
 
         let styleClasses = ['vjs-record-button', 'vjs-control', 'vjs-button',
             'vjs-icon-record-start'];
         styleClasses.forEach((e) => {
-            expect(toggle.hasClass(e)).toEqual(true);
+            expect(toggle.hasClass(e)).toBeTrue();
         });
     });
 
@@ -40,7 +40,7 @@ describe('controls.RecordToggle', function() {
 
         player.one('ready', function() {
             toggle.disable();
-            expect(toggle.enabled_).toEqual(false);
+            expect(toggle.enabled_).toBeFalse();
 
             done();
         });
@@ -49,19 +49,19 @@ describe('controls.RecordToggle', function() {
     it('change appearance when startRecord or stopRecord is triggered', function(done) {
         let toggle = new RecordToggle(player);
 
-        expect(toggle.hasClass('vjs-icon-record-start')).toEqual(true);
+        expect(toggle.hasClass('vjs-icon-record-start')).toBeTrue();
 
         player.one('ready', function() {
             player.trigger('startRecord');
 
-            expect(toggle.hasClass('vjs-icon-record-start')).toEqual(false);
-            expect(toggle.hasClass('vjs-icon-record-stop')).toEqual(true);
+            expect(toggle.hasClass('vjs-icon-record-start')).toBeFalse();
+            expect(toggle.hasClass('vjs-icon-record-stop')).toBeTrue();
             expect(toggle.controlText_).toEqual('Stop');
 
             player.trigger('stopRecord');
 
-            expect(toggle.hasClass('vjs-icon-record-stop')).toEqual(false);
-            expect(toggle.hasClass('vjs-icon-record-start')).toEqual(true);
+            expect(toggle.hasClass('vjs-icon-record-stop')).toBeFalse();
+            expect(toggle.hasClass('vjs-icon-record-start')).toBeTrue();
             expect(toggle.controlText_).toEqual('Record');
 
             done();
@@ -74,7 +74,7 @@ describe('controls.RecordToggle', function() {
         player.one('deviceReady', function() {
             // start
             toggle.trigger('click');
-            expect(player.record().isRecording()).toEqual(true);
+            expect(player.record().isRecording()).toBeTrue();
 
             done();
         });