+++ CHANGES.md
@@ -1,6 +1,16 @@
 videojs-record changelog
 ========================
 
+3.0.0 - unreleased
+------------------
+
+**Backwards-incompatible changes** (when upgrading from a previous version):
+
+- In older versions a `player.recordedData.video` would be available in Chrome
+  when recording audio and video. This is removed and `player.recordedData`
+  is always a `Blob` across all recording types and browsers now (#269)
+
+
 2.4.1 - 2018/08/03
 ------------------